requirements.txt
streamlit==1.32.0
pillow==10.2.0
python-dotenv==1.0.0